Use/don't use field for articles. Rebroadcast => use=0
Add a use field to the articles table. Change the rebroadcast (global
filter) handling to set article_use=0 and not add the enclosure. This
lets us log it the first time we see the article, and also that status
is now in the database.

Change some places to hide/not process article_use=0.