Add per-feed, per-time item limit.
The way this works is that we allow X items over the last Y seconds. So
if you allow 3 items per week (7*86400) [the default], then we first
check how many items were added in the last week *of fetches*. Say
that's 1. Then we'll allow up to 2 more items this fetch. Any more and
we'll roll back this fetch.

We do this by making use of savepoints in the database. Makes rolling
back really easy, and means we can check the limit as we process the
items.

Added a status code to the fetches; this records if a fetch succeeded or
not. Note that its informative only; no decisions are made based on its
value.

More changes to db version 3, which is still NOT FINAL.