Apparently can't use relative URLs; other validation fixes.
Relative URLs are apparently not allowed, at least according to a random
feed validator and Firefox's live bookmarks failing. UGH. The spec
actually says it must be an http URL, but screw that—I'm using https (of
course, it doesn't define that, and you do speak HTTP over https://
URLs—but then again, same with that relative URL).

Also, had guid=0 instead of =false in the feed. Expected FeedPP to
handle that, but apparently not.