Treat empty guid as no guid.
Stack Overflow Blog manages to spit out empty guid elements. I'm not
100% sure, the spec just says "there are no rules for the syntax of a
guid. Aggregators must view them as a string. It's up to the source of
the feed to establish the uniqueness of the string." So probably the
empty string is valid—and really those are all the same article, since
they have the id. But, well, no.