Add start of an fsck routine.
Not really a file system check, but couldn't come up with a better name.
Goal is to do various checks that can't be enforced with SQLite
constraints, or at least not easily.

Currently, it just checks for enclosures which aren't pointed to by any
article_enclosure entry. Artciles_enclosures entries themselves can't be
orphaned due to foreign keys. And it's OK for an article to have no
enclosures.