Add view for usable (*_use = 1) enclosures.
This joins together enclosures and articles, because both of those have
a _use field. Multiple articles can point to the same enclosure; the
enclosure is usable if *any* of the articles have article_use=1.

Also, since we're joining the mess together anyway, and it's useful to
have, two fields derived from articles:

  - min_article_no - MIN(article_no) for all articles linked to the
    enclosure.

  - earliest_article_when - MIN(article_when) for all articles linked to
    the enclosure.

Note the above two may be *different* articles.