DB: Add schema for random items.
Random items have been in Podist for a long time, but they've never been
recorded in the database before. They were just added to the playlists
(m3u files). That was all well and good, until I switched to exporting
RSS feeds which are built entirely from info in the DB.

So first step — add tables to store them in the database.