Random items & their use are now in the DB.
A fair bit of work to do this. Random items are now stored in the
database table "randoms". Each one can be given a weight; the default is
1000. 0 disables an item. 1 makes it extremely unlikely, etc.

Items are automatically inserted when they're found in the RandomMedia
directory. Nothing cleans old items out of the DB, but they're harmless
(except for wasting a tiny bit of disk space). So you can also disable
an item by removing it from RandomMedia.

The use of random items is stored in random_uses. Next to do is to make
the feed generator consult random_uses and put the random items in the
feed.