Fix invalid range for disabled random items.
Disabled (weight=0) random items ended up with a weight range where the
maximum < minimum. Technically, that's the only way to represent it...
so it sort of makes sense, but easier to just remove them from the list
entirely to preserve constraint that min <= max.

Add a separate list of disabled items, so we can still list them. Go
ahead and list them at the bottom so it's easy to find them.