Use iconv instead of recode; handle / specially.
Recode was doing weird things with é — seemed to want to drop a bunch of
characters after it, instead of just turning to e. For example,
"Frédéric" became "Fric". iconv will turn it to Frederic instead. It
also seems to put a ? in place of something it can't convert (instead of
just giving empty). E.g., 日本 becomes ?? instead of (empty).

Also, since / is special on Unix (directory), we shouldn't be putting it
in the suggested title. Change it to -.