Add test data as submodule.
This is to avoid having to use Git-LFS in the main repository, as LFS
doesn't work with simple setups like over ssh, and there isn't a
reasonable lightweight LFS server I could find. So at least this way the
main repository can still be pushed to my internal git server.

*grumble* git-annex doesn't have this problem *grumble*