Change 4ever to mean through now, not last ep.
Previously, 4ever showed the hours/day of a podcast for the period from
its earliest episode through its latest. That means podcasts that put
out a few episodes years ago still inflated the 4ever amount. Really,
that was hours/day during the life of the podcast, so rename that column
"life".

Add a new 4ever column which uses now as the end date, so it's from the
first episode through now.

(Arguably the old 4ever/new lifetime is also a fence-post error, and we
ought to adjust it by adding the average period between episodes to the
denominator, or removing one average episode from the numerator. But
whatever! 66 line queries are long enough.)