Add vorbis encoding for processed audio
Phone (or at least BeyondPod) was having a nightmare from Opus. Every
time it tried to find the length(?) of the file, it took ages,
proportional to length. Also, it'd often cut off the first few seconds
when playing back.

This changed the config a little (use a generic EncodeQuality instead of
Bitrate); update tests to reflect. Added a new test for Vorbis output.

Turns out we need to turn off video (-vn) for vorbis, or attached
pictures become video tracks and things break. Went ahead and added the
option for opus as well, just in case. (Too bad ffmpeg has no good way
to keep them as attachments!)

Hopefully Vorbis works better. Adding MP3 will be pretty trivial, if it
turns out to be needed.