Fix processing MP3 files with both mono and stereo
MP3 can apparently switch between mono and stereo in the middle. That
causes loudnorm to have a very bad day, and Podist to crash. Fix by
ffmpeg-to-ffmpeg pipe, converting to stereo, so that loudnorm never sees
it. Have to do the conversion in a weird way, because ffmpeg insists on
lowering the mono parts by 3dB if you do it the obvious way (-ac 2).
That sort of makes sense (its playing on twice as many speakers), but
not really since playing on both speakers is fully expected. Probably.
Can't be entirely sure what insane mixes expect. [Also, testing shows
various players have interesting ways of handling this — like, say,
just skipping sections of the file.]

Only do this mess for files that require it; it has the downside of
converting mono files to stereo (even in formats that don't require it).
And also loses any cover art. And probably other metadata.

This mess brought to you by Serial season one, new trial update. Which
just had to add 40 mono frames at the end (after 38,598 stereo frames).