Use sqlite's sqldiff to compare databases.
Turns out SQLite has a tool to diff two SQLite databases and show SQL
required to change one to the other. Use this to make sure the upgraded
database matches the current database. Naturally, this found another
failure.

Note that the output has to be filtered because some differences are
expected, in particular the GUID should be different and backup tables
are left behind.

Unfortunately, the SQLite tool does not show the old versions of tables
when there is a schema change; it'll just show "DROP TABLE x" followed
by "CREATE TABLE x ( ——new schema—— )". To work around, when it shows
relevant differences, we re-run it the opposite directing.

CI: Finally, always upload artifacts — not just on success.