Re-create articles on upgrade from v2.
Previously, we just tacked the new columns on at the end. Unfortunately,
that results in sqldiff showing a difference and thus the test fails.
Further, you can't add them to the middle; SQLite only supports adding
at the end.

So do the standard workaround — recreate the table and copy the data.
All tests should now pass!