Start of basic usage test.
Right now it runs the commands and makes sure they don't die (which
already found bugs), but doesn't really verify they do correct things.
There are a bunch of TODO comments about that.

Also, found a bug that isn't a trivial fix so put it in the todo file and
added a todo test for it.