Cache debs as well.
I wish GitLab CI could do multiple caches, but as far as I can find it
can't (and there seem to be at least a few issues open asking for that
support). So switch to one global, and just be careful not to put
anything in there that might change w/o the path changing.

Took four tries, but caching apt-get downloads with a bind mount works.
Also have to remove the Debian Docker image script that auto-nukes the
apt cache.

Haven't found a way to cache the git-lfs downloads. Will be trying again
soon...