GitLab CI is single-core; auto-detect that.
Turns out GitLab CI is single-vCPU, and that's actually documented at
<https://docs.gitlab.com/ee/user/gitlab_com/#shared-runners>. So it
didn't help turning on -j2 there. But it helps a lot when running
locally, probe the core count and enable only when multi-CPU.

That was getting a little to repeaty, so use YAML anchors to re-share
the script. Added a variable to control if we run with coverage testing
or not. (Make sure to quote it, to avoid YAML taking y/n as a boolean).

    “From so simple a typo endless errors most grotesque and most
     tedious have been, and are being, produced.”

(yeah, this took a few tries!)