Test simple playlist generation options.
We turn on and off annoncements, the lead-out, and random media. Confirm
that the DB reflects the expected results. Also gives a good way to be
sure the lead-out is making it to the feed (by adding an absurdly long
lead-out).

Refactored connecting to the DB out of basic-usage.t and into
SystemTesting.pm.