Test music feeds.
Fairly easy to set a few of the existing 8 feeds to music; did so for
feeds 1 and 2. Also set the random chance to 100% and the ratio from
random music feeds to 100%. Check it did it by looking at stderr and
also check to make sure some enclosures from 1/2 were used (a somewhat
nontrivial query).

Upped ridiculous lead-out length to 50 to make sure we don't
accidentally get there.

Made it possible to conect to the DB in read-write mode. Beware of
conflicts with Podist’s access to the DB.