Test various archiving modes; most are TODO.
All the archiving modes other than the one I use aren't implemented.
Should fix that someday, but for now at least its tested (and marked as
a todo in the test).

Unfortunately this means config-variants.t also processes audio, making
it take ~5m. Plan to hack the processor to speed that up, since we
don't actually need the audio processed.