Implement hack in processor to skip processing.
When PODIST_FAST_FAKE_PROCESSOR=1 is set (or any other true value),
attempt to "process" a file by copying it instead of actually processing
it. This is much faster — so much faster, indeed, that it uncovered a
bug in Podist:

One of Podist's debug messages displayed how many times faster than
realtime processing ran. When that was faster than our measurment
resolution, the runtime was 0 leading to division by 0 (and thus a die).
Fixed by testing for that and displaying 'inf'.

Using the new fast mode in config variant test, but not basic-usage.
Want to test the real thing, fully integrated in to Podist, at least
once!

This will slightly reduce coverage, but am not going to write tests to
confirm two simple ifs in a hacked-up optimization used only by the test
suite.