Better log config for test.
Previously, was sending console color codes to stderr, which ultimately
left a bunch of noise in the TAP files. Fix by configuring with a
non-"Simple" log config. Also added the component name since that might
be useful. Log output now looks like this:

    # [   0 T main     ] Log4perl configured and hopefully ready.
    # [   0 T Misc     ] Profile default has no base; done
    # [   0 T Misc     ] Merging profile default into compress
    # [   1 T main     ] Getting terminal size via Term::Size::Any

in the TAP logs, which is much nicer! The reserved width for component
is 10, which is one more than needed for Processor. Also, we now get
really clear multi-line output:

    # [   8 T Storage   ] Deleting processed file /tmp/GOvpnzjYDf/store/playlists/processed/3/002_000_lP7Qrz.mp3
    # [  10 T Storage   ] Deleting processed file /tmp/GOvpnzjYDf/store/playlists/processed/3/003_000_M2t7bl.mp3
    # [  12 F main      ] delete_speech not yet implemented at /src/lib/Podist/Storage.pm line 141.
    #                      at /src/lib/Podist/Storage.pm line 141.
    #                      Podist::Storage::delete_speech(Podist::Storage=HASH(0x55e4a39899f0), 4) called at ./Podist line 1409
    #                      main::__ANON__(4) called at ./Podist line 1427
    #                      main::archive(3) called at ./Podist line 126
    #                      main::go() called at ./Podist line 93

Note how clear it is now that the one starting at time 12 is one
multi-line message, not several messages.

Interestingly, we still get coverage of the simple logging config setup
in Podist because it gets used while writing the example configuration
file.