Add a command to edit existing feeds. Refactor.
Added a new subcommand "editfeed" which edits the various feed
attributes of an existing feed (everything in the "feeds" table). Add a
system test for it as well. Add documentation.

Move the Config->_normalize_time method to Misc and make it not a method
and not private. Add in handling suffix D (days) and W (weeks), both
presuming 24h days. Update tests. Did this to use it for parsing
limit-period.

Clean up a few space-at-eol, since git so loves to highlight them in
bright red...