Add CBR support for speech. Update conf example.
Added an option for mp3-cbr for speech. Doubt it matters except if the
player is *really* broken, you're not likely to be trying to seek
speech. (This hasn't really been tested at all, but it's simple enough
hopefully it works. Plan to test speech options later.)

Also noted the new options in the example config's comments.