Add in a feature to drop a podcast.
This will set all the remaininging unplaylisted episodes from that
podcast to not be used. It also sets the feed to disabled so no new
episodes will be downloaded.

Add a test. Not really much to test, it's just a wrapper around a few
queries. Could be done in one, if you didn't want the log output. (Well,
two — the update to set the feed to enabled=0 is a trivial second).