Test bug: do not don't-use enclosure #1.
Later tests (for example, the one for orphaned enclosures) target
enclosure #1. If we've already marked it don't use, then fsck won't care
about it being messed up, causing test failures.

And since this is all somewhat random (based on which enclosures get
used by a playlist), an annoying random test failure. Pretty sure this
is why f3fe4820 failed on buster (but passed on stretch) and failed
74a2f7fc on stretch. (But neither did locally, despite a few tries, only
on GitLab).