Specify less-secure cipher list; see debian 912759.
Tried to add a podcast with a feed over https, and got a wrong signature
type error. That's apparently from a broken server insiting on SHA-1,
which isn't allowed in the (new) default security level 2. So for now,
downgrade to security level 1. Should revisit this someday, but for now
podcast downloads aren't that security sensitive.