Add time stretching to the audio processor.
This is done with the rubberband filter. Put it before loudnorm because
of course it changes volume slightly.

Found an interesting thing while doing this. That 0 at the end of
Parsed_loudnorm_0 is because it was the first filter in the chain. Now
that we're putting rubberband in front, it can be _1 as well. Take any
number.