Add tests for processor tempo.
This was a little fun since the test I copied from used the processor
again to confirm the generated file has the correct volume. Of course,
when adding the tempo, that means it got tempo scaled again... leading
to a confusing test failure. Avoided by using a second processor w/o
scaling.