Add temporary groups to storage. Make processed names include pid.
Without this, Storage will be copied as part of the fork, and then get
temps added to it. Then when its DESTROY'd, it'd warn about them all
being leaked. Worse, the parent's Storage won't know about the temps, so
will blow up when Podist tries to save them.

Also, for the same fork reasons, add the pid to the generated temp name.
Otherwise, it'll generate the same name for all children...