Important fix: Save last processed item in non-parallel.
Oddly enough the non-parallel "debug" mode in Parallel::ForkManager
exposed a race condition (which is, AFIACT, just a Paralell::ForkManager
bug or API bug): you can't attach a run_on_finish before it finishes,
thus you never get the callback.

Actually, it's worse: it is attached... for the next child. So
processing done with Parallel 0 (the default) made a huge mess.
Thankfully can be fixed by re-processing, (--redo) ... which isn't done
yet, but I'm working on.

Reported this issue at
   https://github.com/dluxhu/perl-parallel-forkmanager/issues/33